✨ Extend onclick JS bridge with style.cssText mutations.
onclick handlers can now also do document.getElementById('id').style.cssText = '...' (and individual style setters once the cssText form is recognized). The assignment scanner was generalized to a (id, prop, value) triple dispatcher, so textContent/innerText (set text) and style.cssText (set inline style, applied via the blitz mutator) share one parser. createAndAppend/addEventListener remain future work (the Boa-native-closure approach is fiddly to get right without an API spike).

Extended offline_onclick_test to assert both: #btn sets #out textContent to 'Hello aris', #sty sets #out style.cssText to 'color:#ff0000' — both pass.

+++ packages/render/src/winit_backend.rs
@@ -1555,7 +1555,7 @@ fn favicon_color(url: &str) -> u32 {
     // Map the hash into HSL space with fixed S/L for a soft, distinct color.
     let hue = (h % 360) as f32;
     let (r, g, b) = hsl_to_rgb(hue, 0.55, 0.55);
-    ((r as u32) << 16) | ((g as u32) << 8) | (b as u32)
+    (r << 16) | (g << 8) | b
 }
 
 fn hsl_to_rgb(h: f32, s: f32, l: f32) -> (u32, u32, u32) {
